feat: 드롭다운 클릭했을 때 모달 나타나는 기능 추가
